P256K1FieldElement: Move integerTo32Bytes from P256K1PubKey
