+++ secp-api/src/main/java/org/bitcoinj/secp/api/P256K1FieldElement.java
@@ -30,4 +30,23 @@ public interface P256K1FieldElement {
     static boolean isInRange(BigInteger x) {
         return x.signum() >= 0 && x.compareTo(Secp256k1.FIELD.getP()) < 0;
     }
+
+    /**
+     * Convert a BigInteger to a fixed-length byte array
+     * @param i an unsigned BigInteger containing a valid Secp256k1 field value
+     * @return a 32-byte, big-endian unsigned integer value
+     */
+    static byte[] integerTo32Bytes(BigInteger i) {
+        // TODO: Check for negative or greater than p?
+        byte[] minBytes = i.toByteArray(); // return minimum, signed bytes
+        if (minBytes.length > 33) throw new IllegalStateException("privKey BigInteger value too large");
+        // Convert from signed, variable length to unsigned, fixed 32-byte length.
+        byte[] result = new byte[32];
+        System.arraycopy(minBytes,                                  // src
+                minBytes.length == 33 ? 1 : 0,                      // src pos (skip sign byte if present)
+                result,                                             // dest
+                minBytes.length == 33 ? 0 : 32 - minBytes.length,   // dest pos
+                minBytes.length == 33 ? 32 : minBytes.length);      // num bytes to copy
+        return result;
+    }
 }
